BugFix: Issue65
I more or less fixed issue 65 with a hack. It is really hard to eliminate an implication and then see if it satisfies the elimination of another implication BUT you in the ImplicationOfAPermission class we can check to see if the other one is of the same type, then go though all of the PermissionSetFromAnnotation objects, convert them to FractionalPermissions, split off and check satisfiability. 

I am a little worried just because I had to redo functionality that definitely already exists, and I hope that we are not silently dropping any constraints.